ENH: Added list_remote_files method and swutched download to use

+++ pysat/instruments/methods/icon.py
@@ -3,11 +3,14 @@
 
 from __future__ import absolute_import, division, print_function
 
+import fnmatch
 import ftplib
 from ftplib import FTP
+import numpy as np
 import os
 
 import pysat
+import pysat._files as pfiles
 
 import logging
 logger = logging.getLogger(__name__)
@@ -74,9 +77,153 @@ ackn_str = ''.join(('This is a data product from the NASA Ionospheric ',
                     '(http://icon.ssl.berkeley.edu/Data).'))
 
 
+def list_remote_files(tag, sat_id, user=None, password=None,
+                      supported_tags=None,
+                      year=None, month=None, day=None,
+                      start=None, stop=None):
+    """Return a Pandas Series of every file for chosen remote data.
+
+    This routine is intended to be used by pysat instrument modules supporting
+    a particular UC-Berkeley SSL dataset related to ICON.
+
+    Parameters
+    -----------
+    tag : (string or NoneType)
+        Denotes type of file to load.  Accepted types are <tag strings>.
+        (default=None)
+    sat_id : (string or NoneType)
+        Specifies the satellite ID for a constellation.  Not used.
+        (default=None)
+    user : (string or NoneType)
+        Username to be passed along to resource with relevant data.
+        (default=None)
+    password : (string or NoneType)
+        User password to be passed along to resource with relevant data.
+        (default=None)
+    start : (dt.datetime or NoneType)
+        Starting time for file list. A None value will start with the first
+        file found.
+        (default=None)
+    stop : (dt.datetime or NoneType)
+        Ending time for the file list.  A None value will stop with the last
+        file found.
+        (default=None)
+
+    Returns
+    --------
+    pandas.Series
+        A Series formatted for the Files class (pysat._files.Files)
+        containing filenames and indexed by date and time
+
+    """
+
+    if (user is not None) or (password is not None):
+        raise ValueError('User account information must not be provided.')
+
+    # connect to CDAWeb default port
+    ftp = FTP('icon-science.ssl.berkeley.edu')
+    # user anonymous, passwd anonymous@
+    ftp.login()
+
+    try:
+        ftp_dict = supported_tags[sat_id][tag]
+    except KeyError:
+        raise ValueError('sat_id/tag name unknown.')
+
+    # naming scheme for files on the CDAWeb server
+    remote_fname = ftp_dict['remote_fname']
+
+    # will hold paths to remote locations
+    remote_years = []
+    remote_days = []
+    temp_dirs = []
+
+    # values of years and days with data
+    years = []
+    days = []
+    yrdoys = []
+
+    # path to highest directory, below which is custom structure
+    # and files
+    # change directory
+    ftp.cwd(ftp_dict['dir'])
+    # get directory contents
+    ftp.dir(temp_dirs.append)
+    # need to parse output to obtain list of paths to years
+    for item in temp_dirs:
+        # parse raw string
+        parsed = item.split(' ')
+        # print(parsed[-1])
+        remote_years.append(ftp_dict['dir'] + '/' + parsed[-1])
+        years.append(parsed[-1])
+
+    # get files under each year, first identify day directories
+    for year, year_int in zip(remote_years, years):
+        ftp.cwd(year)
+        # get list of days
+        temp_dirs = []
+        ftp.dir(temp_dirs.append)
+        for item in temp_dirs:
+            # parse raw string
+            parsed = item.split(' ')
+            # print(parsed[-1])
+            remote_days.append(year + '/' + parsed[-1])
+            days.append(parsed[-1])
+            yrdoys.append(int(year_int)*1000 + int(parsed[-1]))
+
+    # potentially filter here for years and days that are out of bounds
+    yrdoys = np.array(yrdoys)
+
+    # use user supplied date limits or use min and max from observed files
+    if start is None:
+        syrdoy = np.min(yrdoys)
+    else:
+        syr, sdoy = pysat.utils.time.getyrdoy(start)
+        syrdoy = syr * 1000 + sdoy
+
+    if stop is None:
+        eyrdoy = np.max(yrdoys)
+    else:
+        eyr, edoy = pysat.utils.time.getyrdoy(stop)
+        eyrdoy = eyr * 1000 + edoy
+    # apply date filter
+    idx, = np.where((yrdoys >= syrdoy) & (yrdoys <= eyrdoy))
+    remote_days = np.array(remote_days)[idx].tolist()
+
+    # leading path, get any directory elements between the year/doy
+    # and the start of actual files on SSL server
+    leading_dirs = remote_fname.split('/')
+    leading_dirs = leading_dirs[:-1]
+    leading_dirs = '/'.join(leading_dirs)
+    if leading_dirs != '':
+        leading_dirs += '/'
+
+    # get a list of files now that all leading portions determined
+    day_file_list = []
+    for remote_day in remote_days:
+        temp_dirs = []
+        ftp.cwd(remote_day + '/' + leading_dirs)
+        ftp.dir(temp_dirs.append)
+        for item in temp_dirs:
+            # parse raw string
+            parsed = item.split(' ')
+            day_file_list.append(remote_day + '/' + leading_dirs + parsed[-1])
+
+    # we now have a list of all files in the instrument data directories
+    # need to whittle list down to the versions and revisions most appropriate
+    search_dict = pysat._files.construct_searchstring_from_format(remote_fname)
+    search_str = '*/' + search_dict['search_string']
+    remote_files = fnmatch.filter(day_file_list, search_str)
+
+    # pull out date information from the files
+    stored = pfiles.parse_fixed_width_filenames(remote_files, remote_fname)
+    output = pfiles.process_parsed_filenames(stored)
+    # return information, limited to start and stop dates
+    return output[start:stop]
+
+
 def ssl_download(date_array, tag, sat_id, data_path=None,
-                 user=None, password=None, supported_tags=None,
-                 ftp_dir=None):
+                 user=None, password=None, supported_tags=None):
     """Download ICON data from public area of SSL ftp server
 
     Parameters
@@ -110,54 +257,40 @@ def ssl_download(date_array, tag, sat_id, data_path=None,
 
     """
 
-    # connect to CDAWeb default port
-    ftp = FTP('icon-science.ssl.berkeley.edu')
-
-    # user anonymous, passwd anonymous@
-    ftp.login()
-
     try:
         ftp_dict = supported_tags[sat_id][tag]
     except KeyError:
         raise ValueError('sat_id/tag name unknown.')
 
-    # path to relevant file on CDAWeb
-    ftp.cwd(ftp_dict['dir'])
+    # get a list of remote files
+    remote_files = list_remote_files(tag, sat_id, supported_tags=supported_tags,
+                                     start=date_array[0], stop=date_array[-1])
 
-    # naming scheme for files on the CDAWeb server
-    remote_fname = ftp_dict['remote_fname']
+    # connect to CDAWeb default port
+    ftp = FTP('icon-science.ssl.berkeley.edu')
 
-    # naming scheme for local files, should be closely related
-    # to CDAWeb scheme, though directory structures may be reduced
-    # if desired
-    local_fname = ftp_dict['local_fname']
-
+    # user anonymous, passwd anonymous@
+    ftp.login()
     for date in date_array:
-        # format files for specific dates and download location
-        yr, doy = pysat.utils.time.getyrdoy(date)
-        formatted_remote_fname = remote_fname.format(year=yr,
-                                                     doy=doy,
-                                                     month=date.month,
-                                                     day=date.day)
-        formatted_local_fname = local_fname.format(year=yr,
-                                                   doy=doy,
-                                                   month=date.month,
-                                                   day=date.day)
-        saved_local_fname = os.path.join(data_path, formatted_local_fname)
+        if date in remote_files:
+            fname = remote_files[date]
+            # format files for specific dates and download location
+            # yr, doy = pysat.utils.time.getyrdoy(date)
+            saved_local_fname = os.path.join(data_path, fname.split('/')[-1])
 
-        # perform download
-        try:
-            logger.info(' '.join(('Attempting to download file for',
-                                  date.strftime('%x'))))
-            logger.info(formatted_remote_fname)
-            ftp.retrbinary('RETR ' + formatted_remote_fname,
-                           open(saved_local_fname, 'wb').write)
-            logger.info('Finished.')
-        except ftplib.error_perm as exception:
-            if str(exception.args[0]).split(" ", 1)[0] != '550':
-                raise
-            else:
-                os.remove(saved_local_fname)
-                logger.info('File not available for ' + date.strftime('%x'))
+            # perform download
+            try:
+                logger.info(' '.join(('Attempting to download file for',
+                                      date.strftime('%x'))))
+                logger.info(fname)
+                ftp.retrbinary('RETR ' + fname,
+                               open(saved_local_fname, 'wb').write)
+                logger.info('Finished.')
+            except ftplib.error_perm as exception:
+                if str(exception.args[0]).split(" ", 1)[0] != '550':
+                    raise
+                else:
+                    os.remove(saved_local_fname)
+                    logger.info('File not available for ' + date.strftime('%x'))
     ftp.close()
     return
